requirements.txt
azure-ai-documentintelligence==1.0.0b
azure-ai-formrecognizer==3.3.0
azure-common==1.1.28
azure-core==1.30.1
openpyxl==3.0.10
streamlit==1.34.0
